fix warning

+++ src/lib/actions/recipe.action.ts
@@ -1,10 +1,7 @@
 "use server";
 
 import Recipe from "@/lib/models/recipe.model";
-import Category from "../models/category.model";
-import User from "../models/user.model";
 import { connect } from "@/lib/db";
-import { Types } from "mongoose";
 
 type GetRecipesByUserParams = {
   userId: string;
@@ -27,10 +24,9 @@ export async function createRecipe(recipe: any) {
 
 export const getRecipesByUser = async (userId: string) => {
   try {
-    // Ensure the user ID is properly cast to ObjectId if needed
     await connect();
-    const recipes = await Recipe.find({ creator: userId }).exec();
-    return recipes;
+    const recipes = await Recipe.find({ creator: userId }).lean().exec(); // Use lean() to get plain objects
+    return JSON.parse(JSON.stringify(recipes)); // Convert to plain JSON objects
   } catch (error) {
     console.error("Error fetching recipes by user:", error);
     throw error;
@@ -40,42 +36,9 @@ export const getRecipesByUser = async (userId: string) => {
 export const deleteRecipe = async (recipeId: string, userId: string) => {
   try {
     await connect();
-
-    // Ensure the recipe belongs to the user
-    const recipe = await Recipe.findOne({ _id: recipeId, creator: userId });
-
-    if (!recipe) {
-      throw new Error(
-        "Recipe not found or you do not have permission to delete this recipe."
-      );
-    }
-
-    await Recipe.deleteOne({ _id: recipeId });
-    return { message: "Recipe deleted successfully" };
+    await Recipe.deleteOne({ _id: recipeId, creator: userId }).exec();
   } catch (error) {
     console.error("Error deleting recipe:", error);
     throw error;
   }
 };
-
-// export async function updateEvent({ userId, event, path }: UpdateEventParams) {
-//   try {
-//     await connectToDatabase()
-
-//     const eventToUpdate = await Event.findById(event._id)
-//     if (!eventToUpdate || eventToUpdate.organizer.toHexString() !== userId) {
-//       throw new Error('Unauthorized or event not found')
-//     }
-
-//     const updatedEvent = await Event.findByIdAndUpdate(
-//       event._id,
-//       { ...event, category: event.categoryId },
-//       { new: true }
-//     )
-//     revalidatePath(path)
-
-//     return JSON.parse(JSON.stringify(updatedEvent))
-//   } catch (error) {
-//     handleError(error)
-//   }
-// }
